pyproject.toml
[project]
name = "homevm-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "keyring>=25.7.0",
    "paramiko>=4.0.0",
    "pyqt6>=6.10.0",
    "pywinrm>=0.5.0",
    "requests-ntlm>=1.3.0",
]

[dependencies]
PyQt6 = "*"
keyring = "*"
paramiko = "*"
pywinrm = "*"
requests-ntlm = "*"    # WinRM (NTLM) 認証に必要
